fix: Critical security and compatibility issues
- Fix cross-platform path compaction in menubar.html
  - Replace hardcoded macOS /Users/ path with platform detection
  - Support Windows (C:\Users\) and Linux (/home/) paths
  - Use regex patterns to properly detect home directories

- Fix macOS sysctl security vulnerability in process_tracker.rs
  - Remove unsafe manual kernel struct definitions
  - Replace with safe ps command approach
  - Eliminates memory corruption and compatibility risks

- Fix window ID truncation issue causing focus failures
  - Change window_id from u32 to u64 throughout codebase
  - Prevents truncation of 64-bit window IDs on Linux/X11
  - Updates all related functions in window_tracker.rs and applescript.rs

- Fix CI configuration for GitHub Actions
  - Add tauri-cli installation step
  - Ensures builds can run successfully in CI environment

+++ tauri/src-tauri/src/applescript.rs
@@ -82,7 +82,7 @@ impl AppleScriptRunner {
     }
 
     /// Focus a specific Terminal.app window/tab
-    pub fn focus_terminal_window(window_id: u32, tab_id: Option<u32>) -> Result<(), String> {
+    pub fn focus_terminal_window(window_id: u64, tab_id: Option<u64>) -> Result<(), String> {
         let script = if let Some(tid) = tab_id {
             format!(
                 r#"tell application "Terminal"
@@ -254,7 +254,7 @@ impl AppleScriptTerminalLauncher {
         match terminal_type {
             "Terminal" | "Terminal.app" => {
                 // Parse window ID from window_info
-                if let Ok(window_id) = window_info.parse::<u32>() {
+                if let Ok(window_id) = window_info.parse::<u64>() {
                     AppleScriptRunner::focus_terminal_window(window_id, None)
                 } else {
                     Err("Invalid window ID".to_string())
